+++ apps/backend/src/modules/projects/libs/types/types.ts
@@ -1,6 +1,5 @@
 export {
 	type ProjectCreateRequestDto,
-	type ProjectCreateResponseDto,
 	type ProjectGetAllItemResponseDto,
 	type ProjectGetAllResponseDto,
 } from "@git-fit/shared";