+++ apps/frontend/src/libs/components/components.ts
@@ -1,12 +1,13 @@
 export { App } from "./app/app.js";
 export { Avatar } from "./avatar/avatar.js";
+export { Breadcrumbs } from "./breadcrumbs/breadcrumbs.js";
 export { Button } from "./button/button.js";
 export { ConfirmationModal } from "./confirmation-modal/confirmation-modal.js";
 export { Header } from "./header/header.js";
 export { Icon } from "./icon/icon.js";
 export { IconButton } from "./icon-button/icon-button.js";
 export { Input } from "./input/input.js";
-export { Link } from "./link/link.js";
+export { Link, NavLink } from "./link/link.js";
 export { Loader } from "./loader/loader.js";
 export { Modal } from "./modal/modal.js";
 export { PageLayout } from "./page-layout/page-layout.js";