+++ packages/shared/src/modules/projects/libs/types/project-create-response-dto.type.ts
@@ -1,7 +1,0 @@
-type ProjectCreateResponseDto = {
-	description: string;
-	id: number;
-	name: string;
-};
-
-export { type ProjectCreateResponseDto };